Bugfix/fix missing projection in query (#201)
* Fix missing projection in query.

* Adapt types.

+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/dto/MessageStatistics.java
@@ -41,7 +41,7 @@ public class MessageStatistics {
         /**
          * Technical message type.
          */
-        private int numberOfMessages;
+        private long numberOfMessages;
 
         /**
          * The messages statistics grouped by technical message type.
@@ -84,7 +84,7 @@ public class MessageStatistics {
             /**
              * Number of messages.
              */
-            private int numberOfMessages;
+            private long numberOfMessages;
         }
     }
 }
